[ADD] MediaModal component [RENAME] CreateForm component to MediaForm
