+++ watchtower/src/components/Create/Create.tsx
@@ -1,31 +1,15 @@
-import { ReactNode, useContext, useEffect, useRef } from "react";
+import { ReactNode } from "react";
 
 import styles from "./Create.module.css";
 
 import Button from "../../module/Button/Button";
-import CreateForm from "../CreateForm/CreateForm";
 
 import MingcuteAddFill from "../../icons/MingcuteAddFill";
-import { MediaContext } from "../../context/media-context";
+import MediaModal from "../MediaModal/MediaModal";
 
 const Create = (): ReactNode => {
-  const { editingMedia, setEditingMedia } = useContext(MediaContext);
-
-  const dialogRef = useRef<HTMLDialogElement>(null);
-
-  useEffect(() => {
-    if (editingMedia) {
-      dialogRef.current?.showModal();
-    }
-  }, [editingMedia]);
-
   const addButtonClickHandler = (): void => {
     dialogRef.current?.showModal();
-  };
-
-  const closeModal = () => {
-    dialogRef.current?.close();
-    setEditingMedia(null);
   };
 
   return (
@@ -33,10 +17,7 @@ const Create = (): ReactNode => {
       <Button onClick={addButtonClickHandler} shape="circle">
         <MingcuteAddFill />
       </Button>
-      <dialog ref={dialogRef}>
-        {editingMedia && <CreateForm onCancel={closeModal} />}
-        {!editingMedia && <CreateForm onCancel={closeModal} />}
-      </dialog>
+      <MediaModal />
     </div>
   );
 };
